% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_julia_system_requirements.R
\name{check_julia_system_requirements}
\alias{check_julia_system_requirements}
\title{Checks if the used version of Julia and installed verion of a Julia
package are aligned with the requirements in the package's description file.}
\usage{
check_julia_system_requirements(r_pkg)
}
\arguments{
\item{r_pkg}{The name of a package. The specified package's
description file is checked for a required Julia and package version number}
}
\description{
Checks if the used version of Julia and installed verion of a Julia
package are aligned with the requirements in the package's description file.
}
\examples{
check_julia_system_requirements("fwildclusterboot")

}
