% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_julia_packages.R
\name{install_julia_packages}
\alias{install_julia_packages}
\title{Install Julia packages from R}
\usage{
install_julia_packages(packages)
}
\arguments{
\item{Character}{Vector. The Julia packages to be installed}
}
\description{
Install Julia packages from R
}
\examples{
packages <- c("WildBootTests", "FixedEffectModels", "Random")
packages <- c("WildBootTests.jl", "FixedEffectModels.jl", "Random.jl")
install_julia_packages(packages)
}
