% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_set_threads.R
\name{set_julia_ntreads}
\alias{set_julia_ntreads}
\title{Helper function to set the number of threads for use in Julia}
\usage{
set_julia_ntreads()
}
\description{
Helper function to set the number of threads for use in Julia
}
